moved custom gui components into their own package
refactor and cleanup the custom components classes
